Move from old kernel unionfs to FUSE unionfs While FUSE unionfs works in the initrd for '/' as before, there is a significant performance penalty ... up to 300% for some tasks.
As a result, and arguably better, now only '/etc' and '/stat' will be overlaid
resulting in nearly the same overall file system performance as before.
Additionaly we no longer have to setup unionfs in the initrd, but rather in '/etc/rc'.

Most importantly this change will allow a kernel version bump beyound 3.16.x and
does not require kernel patches as Aufs would require.