dhcp6c, add patch to bind to one interface using SO_BINDTODEVICE
It turns out dhcp6c does not work using multiple instances, one for each interface, since the first
IPv6 UDP 546 listen socket will capture all the packets ... that does not work.

While it is possible to only call dhcp6c once for multiple interfaces, we also use ifplugd to monitor
the links, and one instance of dhcp6c for each interface is needed, just as we do with udhcpc.

This is not a general fix, but for Linux and dhcp6c with only one interface specified (as AstLinux does)
the IPv6 UDP 546 listen socket is tied to only one interface.

This patch is quite simple, and is tested to work.